+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed: do not conflict check Shift against itself for last shift check
 * Fixed: link back to Show posts for related Show posts (allow multiple)
 * Fixed: filter next/previous post link for (multiple) related Show posts
+* Fixed: automatic pages conflict where themes filter the_content early
 
 ### 2.3.3.5
 * Fixed: use schedule based on start_day if specified for Schedule view